+++ axum/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+- **added:** Log rejections from built-in extractors with the
+  `axum::rejection=trace` target ([#1890])
 - **fixed:** Fixed performance regression with `Router::nest` introduced in
   0.6.0. `nest` now flattens the routes which performs better ([#1711])
 - **fixed:** Extracting `MatchedPath` in nested handlers now gives the full
@@ -14,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **added:** Implement `Deref` and `DerefMut` for built-in extractors ([#1922])
 
 [#1711]: https://github.com/tokio-rs/axum/pull/1711
+[#1890]: https://github.com/tokio-rs/axum/pull/1890
 [#1922]: https://github.com/tokio-rs/axum/pull/1922
 
 # 0.6.12 (22. March, 2023)